Bugfixes for testing failure
Moved terminal size code to seprate functions based on os
