+++ internal/utils/terminal_size_unix.go
@@ -1,0 +1,17 @@
+//go:build !windows
+// +build !windows
+
+package utils
+
+import (
+	"fmt"
+	"golang.org/x/term"
+	"os"
+)
+
+func getTerminalSize() (int, int, error) {
+	if !term.IsTerminal(int(os.Stdout.Fd())) {
+		return 0, 0, fmt.Errorf("not a terminal")
+	}
+	return term.GetSize(int(os.Stdin.Fd()))
+}
